% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockRectruit_Functions_Likelihood.R
\name{Bevholt}
\alias{Bevholt}
\title{stock recruitment function}
\usage{
Bevholt(ab, ssb)
}
\arguments{
\item{ab}{the model parameters}

\item{ssb}{a vector of ssb}
}
\value{
log recruitment according to model
}
\description{
stock recruitment function
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}
