% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NSR.R
\name{bootscam}
\alias{bootscam}
\title{function name}
\usage{
bootscam(dat, y, n, np, scamfit, sp, nboot)
}
\arguments{
\item{dat}{XXX}

\item{y}{XXX}

\item{n}{XXX}

\item{np}{XXX}

\item{scamfit}{XXX}

\item{sp}{XXX}

\item{nboot}{XXX}
}
\value{
something
}
\description{
function name
}
\author{
Noel Cadigan \email{Noel.Cadigan@mi.mun.ca}
}

