% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc.quantiles}
\alias{calc.quantiles}
\title{Calculate quantiles}
\usage{
calc.quantiles(d, d.det, variable = "variable")
}
\arguments{
\item{d}{xx}

\item{d.det}{xx}

\item{variable}{xx}
}
\description{
xxx
}

