% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDat.R
\name{convertDat}
\alias{convertDat}
\title{plotMSY}
\usage{
convertDat(datfilename, srtype)
}
\arguments{
\item{datfilename}{XXX}

\item{srtype}{XXX}
}
\value{
something
}
\description{
plotMSY
}
\author{
Tim Earl \email{timothy.earl@cefas.co.uk}
}
