\name{convertSumSen}
\alias{convertSumSen}
\title{plotMSY}
\usage{
  convertSumSen(senfilename = NA, indexfilename = NA,
    pfpm = NA, nits = 0, sr = 2, varybiodata = FALSE,
    stockname = "", silent = FALSE, srconstrain = TRUE)
}
\arguments{
  \item{thing}{}
}
\value{
  something
}
\description{
  plotMSY
}
\author{
  Tim Earl \email{timothy.earl@cefas.co.uk}
}

