\name{Eqplot}
\alias{Eqplot}
\title{stock recruitment function}
\usage{
  Eqplot(sim, fit, Blim, Bpa = 1.4 * Blim,
    ymax = c(NA, NA, NA), plot = TRUE)
}
\arguments{
  \item{ab}{the model parameters}

  \item{ssb}{a vector of ssb}
}
\value{
  log recruitment according to model
}
\description{
  stock recruitment function
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

