% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eqsim.R
\name{eqsim_plot_range}
\alias{eqsim_plot_range}
\title{Calculate Fmsy range}
\usage{
eqsim_plot_range(sim, interval = 0.95, type = "median")
}
\arguments{
\item{sim}{XXX}

\item{interval}{XXX}

\item{type}{XXX}
}
\description{
XXX
}

