% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eqsim2.R
\name{eqsim_run2}
\alias{eqsim_run2}
\title{simulates the equilibrium results for a population}
\usage{
eqsim_run2(fit, bio.years = c(2008, 2012), bio.const = FALSE,
  sel.years = c(2008, 2012), sel.const = FALSE, Fscan = c(0.1, 0.2),
  Fcv = 0, Fphi = 0, Blim, Bpa, recruitment.trim = c(3, -3),
  Btrigger = 0, n_years = 100, process.error = TRUE, verbose = TRUE,
  extreme.trim)
}
\arguments{
\item{fit}{A list returned from the function fitModels}

\item{bio.years}{The years to sample maturity, weights and M from}

\item{bio.const}{A flag, if FALSE mean of the biological values from the years selected are used}

\item{sel.years}{The years to sample the selection patterns from}

\item{sel.const}{A flag, if FALSE mean of the selection patterns from the years selected are used}

\item{Fscan}{F values to scan over}

\item{Fcv}{Assessment error in the advisory year}

\item{Fphi}{Autocorrelation in assessment error in the advisory year}

\item{Blim}{This we know}

\item{Bpa}{This we know}

\item{recruitment.trim}{A numeric vector with two log-value clipping the extreme
recruitment values from a continuous lognormal distribution. The values must
be set as c("high","low").}

\item{Btrigger}{If other than 0 (default) the target F applied is reduced by
SSB/Btrigger}

\item{process.error}{Use stochastic recruitment or mean recruitment?  (TRUE = predictive)}

\item{verbose}{Flag, if TRUE (default) indication of the progress of the
simulation is provided in the console. Useful to turn to FALSE when
knitting documents.}

\item{extreme.trim}{Call John Simmonds :-)}

\item{Nrun}{The number of years to run in total (last 50 years from that will be retained)}
}
\description{
XXX
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

