\name{eqsim_summary}
\alias{eqsim_summary}
\title{Summaries the results from eqsim}
\usage{
  eqsim_summary(sim, Blim, Bpa = 1.4 * Blim,
    yield = "landings")
}
\arguments{
  \item{sim}{An object returned from the function
  eqsim_run}

  \item{Blim}{Value for the Blim}

  \item{Bpa}{Value for the Bpa}

  \item{yield}{Character}
}
\description{
  XXX
}
\author{
  Einar Hjorleifsson \email{einar.hjorleifsson@gmail.com}
}

