% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqsr_Buckland.R
\name{eqsr_Buckland}
\alias{eqsr_Buckland}
\title{Stock recruitment fitting}
\usage{
eqsr_Buckland(data, nsamp = 5000, models = c("Ricker", "Segreg", "Bevholt"))
}
\arguments{
\item{data}{must be a data frame containing columns `ssb` and `rec`}

\item{nsamp}{Number of nonparametric bootstrap samples to take from the stock recruitment
data (default is 5000).  If 0 (zero) then only the fits to the
data are returned and no simulations are made.}

\item{models}{A character vector containing stock recruitment models to fit.
User can set any combination of
"Ricker", "Segreg", "Bevholt", "Smooth_hockey".}
}
\description{
Fits one or more stock recruitment relationships to data contained in `data`
}
