% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqsr_fit.R
\name{eqsr_fit}
\alias{eqsr_fit}
\title{Stock recruitment fitting}
\usage{
eqsr_fit(stk, nsamp = 1000, models = c("Ricker", "Segreg", "Bevholt"),
  id.sr = FLCore::name(stk), remove.years = NULL, rshift = 0)
}
\arguments{
\item{stk}{FLStock object}

\item{nsamp}{Number of samples (iterations) to take from the stock recruitment
fit (default is 1000).  If 0 (zero) then only the fits to the
data are returned and no simulations are made.}

\item{models}{A character vector containing stock recruitment models to use
in the model averaging. User can set any combination of
"Ricker", "Segreg", "Bevholt", "Smooth_hockey".}

\item{id.sr}{A character vector specifying an id or name for the stock
recruitment fit being run. The default is to use the slot "name"
in the stk parameter is provided}

\item{remove.years}{A vector specifying the years to remove from the model
fitting.}

\item{rshift}{lag ssb by aditional years (default = 0).  As an example, for
some herring stocks, age 1 (1 winter ring) fish were spawned 2 years
previously, in this case, rshift = 1.}
}
\value{
A list containing the following objects:
\itemize{
  \item `sr.sto` data.frame containing the alpha (a), beta (b), cv and model
        names. The number of rows correspond to the value set of `nsamp` in
        the function call.
  \item `sr.det` The parameters in the stock recruitment model corresponding
        to the "best fit" of any given model.
  \item `pRec` A matrix of predicted recruitment values. The number of rows
        corresponds to the value set in `nsamp` in the function call. The
        number of columns matches with the number of years used in the model
        fitting.
  \item `stk` An FLStock object, same as provided as input by the user.
  \item `rby` A data.frame containing the recruitment (rec), spawning stock
        biomass (ssb) and year used in the fitting of the data.
  \item `id.sr` A string containing run name (taken from the `id.sr` argument)

}
}
\description{
Fits one or more stock recruitment relationship to data containted in an
FLStock object. If more than one stock recruit relationship is provided, the
models are weighted based on smooth AIC weighting (See Buckland et al., 1997).
}
\examples{
\dontrun{
data(icesStocks)
FIT <- eqsr_fit(icesStocks$saiNS,
                nsamp = 1000,
                models = c("Ricker", "Segreg"))

# summary of individual fits
FIT$sr.det
}

}
\references{
Buckland, S.T., K.P. Burnham & N.H. Augustin (1997). Model selection: An integral part of inference.
Biometrics 53, 603-618.
DOI: \href{https://doi.org/10.2307/2533961}{10.2307/2533961}
}
\seealso{
\code{\link{eqsr_plot}} plots a simulation of predictive recruitment
from the fit, and shows a summary of the contributions of each stock
recruitment model to the model average fit.
}
