\name{eqsr_plot2}
\alias{eqsr_plot2}
\title{Summarise predictive distribution of recruitment}
\usage{
  eqsr_plot2(x, n = 5000, ggPlot = TRUE)
}
\arguments{
  \item{x}{A list from eqsr_summary}

  \item{n}{An integer specifying the number of stochastic
  recruitment points to include in the plot}

  \item{ggPlot}{A flag, if TRUE (default) returns a ggplot.
  Base plot (FALSE) not yet implemented.}
}
\value{
  A plot
}
\description{
  Summarise predictive distribution of recruitment
}
\author{
  Einar Hjorleifsson
}

