% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_converter.R
\name{fls2list}
\alias{fls2list}
\title{XXX}
\usage{
fls2list(stk, y, optLand = TRUE)
}
\arguments{
\item{stk}{FLStock object}

\item{y}{year}

\item{optLand}{Boolean, if TRUE (default) then ...}
}
\description{
From Tim
}

