% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockRectruit_Functions_Likelihood.R
\name{initial}
\alias{initial}
\title{Get starting values for models}
\usage{
initial(model, data)
}
\arguments{
\item{model}{the model name, i.e. "Ricker", "Segreg", "Bevholt"}

\item{data}{a data frame with columns rec and ssb}
}
\value{
vector of starting values
}
\description{
Starting values for hockey stick models
}
