\name{llik}
\alias{llik}
\title{the log-likelihood of the rectuit function}
\usage{
  llik(param, data, model, logpar = FALSE)
}
\arguments{
  \item{param}{the model parameters}

  \item{data}{the rec and ssb data}

  \item{model}{the stock recruit model to use}

  \item{logpar}{are the parameters on the log scale}
}
\value{
  the log-likelihood
}
\description{
  the log-likelihood of the rectuit function
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

