\name{LLmodplot}
\alias{LLmodplot}
\title{plots the ordered posterior densities accross MCMC iterations for each model considered in \code{fitModels}}
\usage{
  LLmodplot(fit)
}
\arguments{
  \item{fit}{an fitted MCMC returned from \code{fitModels}}

  \item{fit}{an fitted MCMC returned from \code{fitModels}}
}
\value{
  NULL produces a plot

  NULL produces one or several plots
}
\description{
  plots the ordered posterior densities accross MCMC
  iterations for each model considered in \code{fitModels}
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}

  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

