% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLplot_LLmodplot.R
\name{LLplot}
\alias{LLplot}
\title{plots the ordered posterior densities accross MCMC iterations for each model considered in \code{fitModels}}
\usage{
LLplot(fit)
}
\arguments{
\item{fit}{an fitted MCMC returned from \code{fitModels}}
}
\value{
NULL produces a plot
}
\description{
plots the ordered posterior densities accross MCMC iterations for each model considered in \code{fitModels}
}
\author{
Colin Millar \email{colin.millar@ices.dk}
}
