% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mydf}
\alias{mydf}
\title{mydf}
\usage{
mydf(rho, data, nknots, my.df)
}
\arguments{
\item{rho}{smoothing parameter}

\item{data}{XXX}

\item{nknots}{XXX}

\item{my.df}{XXX}
}
\value{
something
}
\description{
mydf
}
\author{
Noel Cadigan \email{Noel.Cadigan@mi.mun.ca}
}

