% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NSR.R
\name{mygcv}
\alias{mygcv}
\title{mygcv}
\usage{
mygcv(dat, rho, n)
}
\arguments{
\item{dat}{XXX}

\item{rho}{a parameter}

\item{n}{XXX}
}
\value{
something
}
\description{
mygcv
}
\author{
Noel Cadigan \email{Noel.Cadigan@mi.mun.ca}
}

