% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateRefPts.R, R/plotMSY.r
\name{paste0}
\alias{paste0}
\title{A function only available for R 2.15.1}
\usage{
paste0(..., sep = "")

paste0(..., sep = "")
}
\arguments{
\item{...}{whatever}

\item{sep}{XXX}
}
\value{
result of paste with sep = ""
}
\description{
for back compatibility

XXX
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

