% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_basePlots.R
\name{plot_type1}
\alias{plot_type1}
\title{Plot type 1}
\usage{
plot_type1(p, d, xlim, ylim)
}
\arguments{
\item{p}{XXX}

\item{d}{XXX}

\item{xlim}{XXX}

\item{ylim}{XXX}
}
\description{
Plot type 1
}

