% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_basePlots.R
\name{plot_type2}
\alias{plot_type2}
\title{Plot type 2}
\usage{
plot_type2(sto, det, n, xlim, ylim)
}
\arguments{
\item{sto}{XXX}

\item{det}{XXX}

\item{n}{XX}

\item{xlim}{XXX}

\item{ylim}{XXX}
}
\description{
Plot type 2
}

