\name{read.Lowestoft}
\alias{read.Lowestoft}
\title{plotMSY}
\usage{
read.Lowestoft(filename = NA, index = 1:10, expand = TRUE,
  silent = FALSE)
}
\arguments{
  \item{filename}{XXX}

  \item{index}{XXX}

  \item{expand}{XXX}

  \item{silent}{XXX}
}
\value{
something
}
\description{
plotMSY
}
\author{
Tim Earl \email{timothy.earl@cefas.co.uk}
}

