% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.Stockfile.R
\name{read.Stockfile}
\alias{read.Stockfile}
\title{plotMSY}
\usage{
read.Stockfile(filename, expand = FALSE, index = NA, silent = FALSE)
}
\arguments{
\item{filename}{XXX}

\item{expand}{XXX}

\item{index}{XXX}

\item{silent}{XXX}
}
\value{
something
}
\description{
plotMSY
}
\author{
Tim Earl \email{timothy.earl@cefas.co.uk}
}
