% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockRectruit_Functions_Likelihood.R
\name{segreg2}
\alias{segreg2}
\title{stock recruitment function}
\usage{
segreg2(ab, ssb)
}
\arguments{
\item{ab}{the model parameters}

\item{ssb}{a vector of ssb}
}
\value{
log recruitment according to model
}
\description{
stock recruitment function
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}
