\docType{package}
\name{SimEq-package}
\alias{SimEq}
\alias{SimEq-package}
\title{Estimation of equilibrium reference points for fish stocks}
\description{
  Estimation of equilibrium reference points for fish
  stocks
}
\details{
  \tabular{ll}{ Package: \tab SimEq \cr Type: \tab Package
  \cr Version: \tab 0.1.1 \cr Date: \tab 2013-01-24 \cr
  License: \tab GPL (>= 2) \cr LazyLoad: \tab yes \cr }
}
\examples{
x <- 1
cat(x)
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}

  John Simmonds \email{colin.millar@jrc.ec.europa.eu}

  Noel Cadigan \email{colin.millar@jrc.ec.europa.eu}
}
\references{
  Simmonds, Cadigan
}
\seealso{
  \code{\link{paste}}
}
\keyword{package}

