% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockRectruit_Functions_Likelihood.R
\name{Smooth_hockey}
\alias{Smooth_hockey}
\title{stock recruitment function}
\usage{
Smooth_hockey(ab, ssb, gamma = 0.1)
}
\arguments{
\item{ab}{the model parameters}

\item{ssb}{a vector of ssb}

\item{gamma}{a smoother parameter}
}
\value{
log recruitment according to model
}
\description{
stock recruitment function
}
