% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_read.R
\name{srmsymc_bevholt}
\alias{srmsymc_bevholt}
\title{srmsymc_bevholt}
\usage{
srmsymc_bevholt(ssb, alpha, beta)
}
\arguments{
\item{ssb}{spawning stock biomass}

\item{alpha}{alpha}

\item{beta}{beta}
}
\description{
srmsymc_bevholt
}

