% Generated by roxygen2 (4.0.1): do not edit by hand
\name{srmsymc_cat_age}
\alias{srmsymc_cat_age}
\title{Write age.dat to disk}
\usage{
srmsymc_cat_age(filename_age, n_fleets, pf, pm, sel, sel_cv, w, w_cv, bio,
  bio_cv)
}
\arguments{
\item{filename_age}{XXX}

\item{n_fleets}{XXX}

\item{pf}{XXX}

\item{pm}{XXX}

\item{sel}{XXX}

\item{sel_cv}{XXX}

\item{w}{XXX}

\item{w_cv}{XXX}

\item{bio}{XXX}

\item{bio_cv}{XXX}
}
\description{
XXX
}
\author{
Einar Hjorleifsson
}

