% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_read.R
\name{srmsymc_combine_parameters}
\alias{srmsymc_combine_parameters}
\title{Combine the parameters from the three sr-functions}
\usage{
srmsymc_combine_parameters(trimming = NA)
}
\arguments{
\item{trimming}{XXX}
}
\description{
Calculates weights
}

