\name{srmsymc_combineParameterFiles}
\alias{srmsymc_combineParameterFiles}
\title{Combine the parameters from the three sr-functions}
\usage{
  srmsymc_combineParameterFiles(srweights = NA)
}
\arguments{
  \item{srweights}{a vector of length 3 indicating the
  weights for each srmodel. If missing then equal weights
  applied to all models.}
}
\description{
  XXX
}

