% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_run.R
\name{srmsymc_compile}
\alias{srmsymc_compile}
\title{Compile the srmsymc ADMB code}
\usage{
srmsymc_compile(windose = TRUE, compile = TRUE)
}
\arguments{
\item{windose}{Whether in windows or not}

\item{compile}{Compile binaries from source. TRUE requires ADMB to be installed, FALSE only works on Windows}
}
\description{
The \code{plotMSY} approach uses a ADModel Builder code. ...
}

