% Generated by roxygen2 (4.0.1): do not edit by hand
\name{srmsymc_get_modelname}
\alias{srmsymc_get_modelname}
\title{Get the name of the recruitment function}
\usage{
srmsymc_get_modelname(srno)
}
\arguments{
\item{srno}{The recruitment model number}
}
\description{
This is used in some plotting routines
}

