% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_basePlots.R
\name{srmsymc_plot_distribution}
\alias{srmsymc_plot_distribution}
\title{XXX}
\usage{
srmsymc_plot_distribution(variable = "fmsy", weights = "automatic")
}
\arguments{
\item{variable}{XXX}

\item{weights}{XXX}
}
\description{
XXX
}

