% Generated by roxygen2 (4.0.1): do not edit by hand
\name{srmsymc_plot_distribution_fmort}
\alias{srmsymc_plot_distribution_fmort}
\title{XXX}
\usage{
srmsymc_plot_distribution_fmort(variable = "fmsy", weights = "automatic")
}
\arguments{
\item{variable}{XXX}

\item{weights}{XXX}
}
\description{
XXX
}

