% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_basePlots.R
\name{srmsymc_plotcomposit_ssbr}
\alias{srmsymc_plotcomposit_ssbr}
\title{XXX}
\usage{
srmsymc_plotcomposit_ssbr(rby, n = 40)
}
\arguments{
\item{rby}{XXX}

\item{n}{XXX}
}
\description{
XXX
}

