\name{srmsymc_plotComposite}
\alias{srmsymc_plotComposite}
\title{XXX}
\usage{
  srmsymc_plotComposite(path = "ricker", rby, n = 40, Fbar,
    Fpa, Flim)
}
\arguments{
  \item{path}{list}

  \item{rby}{data.frame with stock summary data}

  \item{n}{XXX}

  \item{Fbar}{XXX}

  \item{Fpa}{XXX}

  \item{Flim}{XXX}
}
\description{
  XXX
}

