\name{srmsymc_plotDistribution}
\alias{srmsymc_plotDistribution}
\title{XXX}
\usage{
  srmsymc_plotDistribution(path, variable = "fmsy",
    srweights = NA, stockname = "The stock")
}
\arguments{
  \item{path}{XXX}

  \item{variable}{XXX}

  \item{srweights}{XXX}

  \item{stockname}{XXX}
}
\description{
  XXX
}

