% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_basePlots.R
\name{srmsymc_plotPar}
\alias{srmsymc_plotPar}
\title{XXX}
\usage{
srmsymc_plotPar(path = "ricker", Fbar, Fpa, Flim)
}
\arguments{
\item{path}{list}

\item{Fbar}{XXX}

\item{Fpa}{XXX}

\item{Flim}{XXX}
}
\description{
XXX
}

