% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_basePlots.R
\name{srmsymc_plotSSB}
\alias{srmsymc_plotSSB}
\title{XXX}
\usage{
srmsymc_plotSSB(path = "ricker", rby, n = 100, hair = FALSE, ylim)
}
\arguments{
\item{path}{list}

\item{rby}{data.frame with stock summary data}

\item{n}{XXX}

\item{hair}{XXX}

\item{ylim}{XXX}
}
\description{
XXX
}

