% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_basePlots.R
\name{srmsymc_plotSSBR}
\alias{srmsymc_plotSSBR}
\title{XXX}
\usage{
srmsymc_plotSSBR(path = "ricker", rby, n = 30, hair = FALSE)
}
\arguments{
\item{path}{list}

\item{rby}{data.frame with stock summary data}

\item{n}{XXX}

\item{hair}{XXX}
}
\description{
XXX
}

