% Generated by roxygen2 (4.0.1): do not edit by hand
\name{srmsymc_plotYield}
\alias{srmsymc_plotYield}
\title{XXX}
\usage{
srmsymc_plotYield(path = "ricker", rby, ylim, n = 100, hair = FALSE)
}
\arguments{
\item{path}{list}

\item{rby}{data.frame with stock summary data}

\item{ylim}{XXX}

\item{n}{XXX}

\item{hair}{XXX}
}
\description{
XXX
}

