% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_read.R
\name{srmsymc_read}
\alias{srmsymc_read}
\title{Read all output files from srmsymc program}
\usage{
srmsymc_read(path, ...)
}
\arguments{
\item{path}{x}

\item{...}{addition arguements}
}
\description{
xxx
}

