% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_read.R
\name{srmsymc_read_setup}
\alias{srmsymc_read_setup}
\title{Read the srmsymc setup}
\usage{
srmsymc_read_setup(path)
}
\arguments{
\item{path}{The path to the directory that stores the srmsymc results}
}
\description{
Reads the setup from the srmsymc.dat file
}

