% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_read.R
\name{srmsymc_read_yield}
\alias{srmsymc_read_yield}
\title{Read the yield estimates the srmsymc program}
\usage{
srmsymc_read_yield(path, file = "simpary.dat", trim = TRUE,
  longformat = TRUE, doPlot = TRUE)
}
\arguments{
\item{path}{character containing the path to the file. If missing it is
assumed that the file is in the current working directory.}

\item{file}{the file name, should not be anything else than \code{simpary.dat}}

\item{trim}{boolean. If TRUE (default) come cleaning (NEED TO SPECIFY) is done.}

\item{longformat}{boolean. If TRUE (default) return a long format.}

\item{doPlot}{boolean. If TRUE (default) returns a ggplot}
}
\description{
Reads the file \code{simpary.dat} generate by the \code{srmsymc}.
In the \code{simpary.dat} the first row is the fishing mortality,
the second row is the deterministic estimates of yield and the remaining
rows represents the catch for each iterations.
Returns a list containing two data.frames, one containing the deterministic
estimate, the other containing the stochastic simulations.
}

