% Generated by roxygen2 (4.0.1): do not edit by hand
\name{srmsymc_read_ypr}
\alias{srmsymc_read_ypr}
\title{Read the yield per recruit estimates the srmsymc program}
\usage{
srmsymc_read_ypr(file = "simparypr.dat", trim = TRUE, longformat = TRUE,
  doPlot = TRUE)
}
\arguments{
\item{file}{the file name, should not be anything else than \code{simparypr.dat}}

\item{trim}{boolean. If TRUE (default) come cleaning (NEED TO SPECIFY) is done.}

\item{longformat}{boolean. If TRUE (default) return a long format.}

\item{doPlot}{boolean. If TRUE (default) returns a ggplot}
}
\description{
Reads the file \code{simparypr.dat} generate by the \code{srmsymc}.
In the \code{simparypr.dat} the first row is the fishing mortality,
the second row is the deterministic estimates of ssb and the remaining
rows represents the catch for each iterations.
Returns a list containing two data.frames, one containing the deterministic
estimate, the other containing the stochastic simulations.
}

