% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_read.R
\name{srmsymc_readRecModel}
\alias{srmsymc_readRecModel}
\title{Recruitment model}
\usage{
srmsymc_readRecModel(path)
}
\arguments{
\item{path}{XXX}
}
\description{
Read recruitment model from the srmsymc.dat file
}

