% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_read.R
\name{srmsymc_segreg}
\alias{srmsymc_segreg}
\title{srmsymc_segreg}
\usage{
srmsymc_segreg(ssb, alpha, beta)
}
\arguments{
\item{ssb}{spawning stock biomass}

\item{alpha}{alpha}

\item{beta}{beta}
}
\description{
srmsymc_segreg
}

