\name{srmsymc_srweights}
\alias{srmsymc_srweights}
\title{Estimating weights for each recruitment function}
\usage{
  srmsymc_srweights(srautoweights = TRUE, trimming = NA,
    nits = 100)
}
\arguments{
  \item{srautoweights}{XXX}

  \item{trimming}{XXX}

  \item{nits}{Number of iterations}
}
\description{
  XXX
}

