\name{SRplot}
\alias{SRplot}
\title{plot simulated predictive distrubution of recruitment}
\usage{
  SRplot(fit)
}
\arguments{
  \item{fit}{an fitted MCMC returned from \code{fitModels}}
}
\value{
  NULL produces a plot
}
\description{
  plot simulated predictive distrubution of recruitment
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

