\name{updateparam}
\alias{updateparam}
\title{performs a single}
\usage{
  updateparam(param, data, llikhood, delta, model)
}
\arguments{
  \item{param}{the y location of the bubbles}

  \item{data}{a scaling factor}

  \item{llikhood}{extra arguments to plot and points}

  \item{delta}{extra arguments to plot and points}

  \item{model}{extra arguments to plot and points}
}
\value{
  a pointer to the environment in which summaries of the
  data reside
}
\description{
  performs a single
}
\note{
  \code{extractData} is intended to be used internally
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

