\name{bevholt}
\alias{bevholt}
\title{stock recruitment function}
\usage{
  bevholt(ab, ssb)
}
\arguments{
  \item{ab}{the model parameters}

  \item{ssb}{a vector of ssb}
}
\value{
  log recruitment according to model
}
\description{
  stock recruitment function
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

