% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/myMH.R
\name{bevholt2}
\alias{bevholt2}
\title{stock recruitment function}
\usage{
bevholt2(ab, ssb)
}
\arguments{
\item{ab}{the model parameters}

\item{ssb}{a vector of ssb}
}
\value{
log recruitment according to model
}
\description{
stock recruitment function
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

