% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateRefPts.R
\name{Eqplot}
\alias{Eqplot}
\title{Plot summary of the equilibrium simulation}
\usage{
Eqplot(sim, fit, Blim, Bpa = 1.4 * Blim, ymax = c(NA, NA, NA),
  plot = TRUE, yield = "landings", extreme.trim)
}
\arguments{
\item{sim}{A list obtained from function \code{EqSim}}

\item{fit}{A list obtained from function \code{fitModels}}

\item{Blim}{A value for Blim}

\item{Bpa}{A value for Bpa. If not specified Bpa is set to 1.4 Blim}

\item{ymax}{A vector of length three indicating the maximum values for
each of the first three panel plots}

\item{plot}{Flag, if TRUE (default) returns both values and plot}

\item{yield}{Character vector of length 1, either "landings" (default) or "catch"}

\item{extreme.trim}{Call John Simmonds :-)}
}
\description{
Plot summary of the equilibrium simulation
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

