\name{eqsim_plot}
\alias{eqsim_plot}
\title{Plots of the results from eqsim}
\usage{
  eqsim_plot(x, Scale = 1000, plotit = TRUE)
}
\arguments{
  \item{x}{An object returned from the function
  eqsim_summary}

  \item{Scale}{A value, the scaling on the yaxis}

  \item{plotit}{Boolean, if TRUE (default) returns a plot}
}
\description{
  XXX
}
\author{
  Einar Hjorleifsson \email{einar.hjorleifsson@gmail.com}
}

