% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqsim.R
\name{eqsim_plot}
\alias{eqsim_plot}
\title{Plots of the results from eqsim}
\usage{
eqsim_plot(sim, ymax.multiplier = 1.2, catch = TRUE)
}
\arguments{
\item{sim}{An object returned from the function eqsim_run}

\item{ymax.multiplier}{A value that acts as a multiplier of the maximum observed
variable being plotted. E.g. 1.2 means that for each of the three panels a, b and c
the ymax is set to 1.2 of the maximum observed recruitment, spawning stock biomass
and yield (catch or landings, depending on user input.}

\item{catch}{Boolean, if TRUE (default) returns a plot based on catch. If false
returns a plot based on landings.}
}
\description{
XXX
}
\author{
Einar Hjorleifsson \email{einar.hjorleifsson@gmail.com}
}

