% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqsim.R
\name{eqsim_run}
\alias{eqsim_run}
\title{simulates the equilibrium results for a population}
\usage{
eqsim_run(fit, bio.years = c(2008, 2012), bio.const = FALSE,
  sel.years = c(2008, 2012), sel.const = FALSE, Fscan = seq(0, 1, len =
  20), Fcv = 0, Fphi = 0, SSBcv = 0, rhologRec = FALSE, Blim, Bpa,
  recruitment.trim = c(3, -3), Btrigger = 0, Nrun = 200,
  process.error = TRUE, verbose = TRUE, extreme.trim)
}
\arguments{
\item{fit}{A list returned from the function fitModels}

\item{bio.years}{The years to sample maturity, weights and M from}

\item{bio.const}{A flag, if FALSE mean of the biological values from the years selected are used}

\item{sel.years}{The years to sample the selection patterns from}

\item{sel.const}{A flag, if FALSE mean of the selection patterns from the years selected are used}

\item{Fscan}{F values to scan over}

\item{Fcv}{Assessment error in the advisory year}

\item{Fphi}{Autocorrelation in assessment error in the advisory year}

\item{SSBcv}{Spawning stock biomass error in the advisory year}

\item{rhologRec}{A flag for recruitment autocorrelation. If FALSE (default) 
then not applied.}

\item{Blim}{This we know}

\item{Bpa}{This we know}

\item{recruitment.trim}{A numeric vector with two log-value clipping the extreme
recruitment values from a continuous lognormal distribution. The values must
be set as c("high","low").}

\item{Btrigger}{If other than 0 (default) the target F applied is reduced by
SSB/Btrigger}

\item{Nrun}{The number of years to run in total (last 50 years from that will be retained)}

\item{process.error}{Use stochastic recruitment or mean recruitment?  (TRUE = predictive)}

\item{verbose}{Flag, if TRUE (default) indication of the progress of the
simulation is provided in the console. Useful to turn to FALSE when 
knitting documents.}

\item{extreme.trim}{Call John Simmonds :-)}
}
\description{
XXX
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

