% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqsr_Buckland.R
\name{eqsr_Buckland}
\alias{eqsr_Buckland}
\title{stock recruitment function}
\usage{
eqsr_Buckland(data, nsamp = 5000, models = c("Ricker", "Segreg", "Bevholt"),
  ...)
}
\arguments{
\item{data}{data.frame containing stock recruitment data}

\item{nsamp}{Number of samples}

\item{models}{A character vector}

\item{...}{Additional arguements}
}
\value{
log recruitment according to model
}
\description{
stock recruitment function
}
\author{
Colin Millar \email{colin.millar@ices.dk}
}
