% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eqsr.R
\name{eqsr_plot}
\alias{eqsr_plot}
\title{plot simulated predictive distribution of recruitment}
\usage{
eqsr_plot(fit, n = 5000, x.mult = 1.1, y.mult = 1.4, ggPlot = FALSE,
  Scale = 1)
}
\arguments{
\item{fit}{an fitted MCMC returned from \code{eqsr_fit}}

\item{n}{Number of random recruitment draws to plot}

\item{x.mult}{max.y (ssb) as a multiplier of maximum observed ssb}

\item{y.mult}{max.x (rec) as a multiplier of maismum observed rec}

\item{ggPlot}{Flag, if FALSE (default) plot base graphics, if true
do a ggplot}

\item{Scale}{Numeric value for scaling varibles in plot.}
}
\value{
NULL produces a plot
}
\description{
plot simulated predictive distribution of recruitment
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

