% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqsr_plot.R
\name{eqsr_plot}
\alias{eqsr_plot}
\title{Plot Simulated Predictive Distribution of Recruitment}
\usage{
eqsr_plot(
  fit,
  n = 20000,
  x.mult = 1.1,
  y.mult = 1.4,
  ggPlot = FALSE,
  Scale = 1
)
}
\arguments{
\item{fit}{an fitted stock recruit model returned from \code{eqsr_fit}}

\item{n}{Number of random recruitment draws to plot}

\item{x.mult}{max value for the y axis (ssb) as a multiplier of maximum
observed ssb}

\item{y.mult}{max value for the x axis (rec) as a multiplier of maismum
observed rec}

\item{ggPlot}{Flag, if FALSE (default) plot using base graphics, if TRUE
do a ggplot}

\item{Scale}{Numeric value for scaling varibles in plot.}
}
\value{
NULL produces a plot
}
\description{
Plot Simulated Predictive Distribution of Recruitment
}
\examples{

\dontrun{
data(icesStocks)
FIT <- eqsr_fit(icesStocks$saiNS,
                nsamp = 1000,
                models = c("Ricker", "Segreg"))

eqsr_plot(FIT, n = 20000)

# Scale argument only available for ggPlot = TRUE
eqsr_plot(FIT, n = 20000, ggPlot = TRUE, Scale = 1000)
}

}
\seealso{
\code{\link{eqsr_fit}} Fits several stock recruitment models to a data set
and calculates the proportion contribution of each model based on a bootstrap
model averaging procedure.
}
