% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModelsSimmonds.R
\name{fitModelsSimmonds}
\alias{fitModelsSimmonds}
\title{stock recruitment function}
\usage{
fitModelsSimmonds(data, runid, delta = 1.3, nburn = 10000, nsamp = 5000,
  models = c("ricker", "segreg", "bevholt"), ...)
}
\arguments{
\item{data}{data.frame containing stock recruitment data}

\item{runid}{A character vector}

\item{delta}{A value}

\item{nburn}{An integer, specifying the burn-in period}

\item{nsamp}{An integer, specifying the number of samples}

\item{models}{A character vector specifying stock-recruitment models}

\item{...}{Additional arguments}
}
\value{
log recruitment according to model
}
\description{
stock recruitment function
}
\author{
Colin Millar \email{colin.millar@ices.dk}
}
