% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateRefPts.R
\name{initial}
\alias{initial}
\title{Get starting values for models}
\usage{
initial(model, data)
}
\arguments{
\item{model}{XXX}

\item{data}{XXX}
}
\value{
vector of starting values
}
\description{
a quick fix!!
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

