% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateRefPts.R
\name{LLmodplot}
\alias{LLmodplot}
\title{plots the fits for each model considered in \code{fitModels}}
\usage{
LLmodplot(fit)
}
\arguments{
\item{fit}{an fitted MCMC returned from \code{fitModels}}
}
\value{
NULL produces one or several plots
}
\description{
plots the fits for each model considered in \code{fitModels}
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

