% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateRefPts.R
\name{loader}
\alias{loader}
\title{Progress function}
\usage{
loader(p)
}
\arguments{
\item{p}{Value}
}
\description{
Progress function
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

