% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loader}
\alias{loader}
\title{Progress function}
\usage{
loader(p)
}
\arguments{
\item{p}{Value}
}
\description{
plots the prgress of an iterative procedure using "[=>   ]", "[==> ]", etc.
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}
