% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msy-package.R
\docType{package}
\name{msy-package}
\alias{msy-package}
\alias{msy}
\title{Estimation of equilibrium reference points for fish stocks}
\description{
Estimation of equilibrium reference points for fish stocks
}
\details{
\tabular{ll}{
Package: \tab msy \cr
Type: \tab Package \cr
Version: \tab 0.1.1 \cr
Date: \tab 2013-10-03 \cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes \cr
}
}
\references{
Simmonds, Cadigan, Steve Smith
}
\author{
Colin Millar \email{colin.millar@ices.dk}

Noel Cadigan \email{Noel.Cadigan@mi.mun.ca}

Tim Earl \email{timothy.earl@cefas.co.uk}

Jose D'Oliveira \email{jose.deoliveira@cefas.co.uk}
}
\keyword{package}
