\name{srmsymc_calc_srweights}
\alias{srmsymc_calc_srweights}
\title{Estimating weights for each recruitment function}
\usage{
  srmsymc_calc_srweights(trimming = NA, srweights = NA)
}
\arguments{
  \item{trimming}{proportion of the simulations to trim
  before taking harmonic mean.}

  \item{srweights}{XXX}
}
\description{
  XXX
}

