\name{srmsymc_compile}
\alias{srmsymc_compile}
\title{Compile the srmsymc ADMB code}
\usage{
  srmsymc_compile(windose = TRUE, compile = TRUE)
}
\arguments{
  \item{windose}{Whether in windows or not}

  \item{compile}{Compile binaries from source. TRUE
  requires ADMB to be installed, FALSE only works on
  Windows}
}
\description{
  The \code{plotMSY} approach uses a ADModel Builder code.
  ...
}

