% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_basePlots.R
\name{srmsymc_get_modelname}
\alias{srmsymc_get_modelname}
\title{Get the name of the recruitment function}
\usage{
srmsymc_get_modelname(srno)
}
\arguments{
\item{srno}{The recruitment model number}
}
\description{
This is used in some plotting routines
}

