% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srmsymc_basePlots.R
\name{srmsymc_plotcomposit_yield}
\alias{srmsymc_plotcomposit_yield}
\title{XXX}
\usage{
srmsymc_plotcomposit_yield(path = "ricker", ylim, rby, n = 40, Fbar, Fpa,
  Flim)
}
\arguments{
\item{path}{list}

\item{ylim}{XXX}

\item{rby}{data.frame with stock summary data}

\item{n}{XXX}

\item{Fbar}{XXX}

\item{Fpa}{XXX}

\item{Flim}{XXX}
}
\description{
XXX
}

