% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRplot.R
\name{SRplot}
\alias{SRplot}
\title{plot simulated predictive distribution of recruitment}
\usage{
SRplot(fit, n = 5000)
}
\arguments{
\item{fit}{an fitted MCMC returned from \code{fitModels}}

\item{n}{Number of random recruitment draws to plot}
}
\value{
NULL produces a plot
}
\description{
plot simulated predictive distribution of recruitment
}
\author{
Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}
